Extend hack to support other common corruption
diff3 seems to output strange things in a couple of different forms when
files don't end in a newline.  The fact that the files I'm working with
have DOS line endings may have something to do with this.  The original
version handled the following corruption, seen when running the code on
my FreeBSD machine:

    <<<<<<< BASE
    LastLine=======
    LastLine
    >>>>>>> YOURS

When running the code on a Windows machine, the following corruption was
observed:

    <<<<<<< BASE
    LastLine
    =======
    LastLine>>>>>>> YOURS

This patch allows the script to address either case.