Resolve 2-way conflicts automatically
This script was originally written to handle 3-way conflicts by
presenting the user with a choice of which version to use.  Sometimes,
however, diff3 outputs conflict markers for only two versions.  This is
the case when both YOURS and MINE have diverged from BASE with the same
change.  In our use-case we can assume that the new change is correct,
so we just accept it automatically.